# docs/swagger.yaml
definitions:
  types.Booking:
    properties:
      car_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      start_date:
        type: string
      total:
        type: number
      user_id:
        type: integer
    type: object
  types.Car:
    properties:
      color:
        type: string
      company_id:
        description: ID of the company that owns the car
        type: integer
      created_at:
        type: string
      id:
        type: integer
      make:
        description: Make (e.g., Toyota, Ford)
        type: string
      model:
        description: Model (e.g., Corolla, Mustang)
        type: string
      price_per_day:
        type: number
      registration_no:
        description: Car registration number
        type: string
      updated_at:
        description: Last updated timestamp
        type: string
      year:
        description: Year of manufacture
        type: integer
    type: object
...
  types.CreateBookingPayload:
    properties:
      car_id:
        type: integer
      end_date:
        type: string
      start_date:
        type: string
    required:
    - car_id
    - end_date
    - start_date
    type: object
  types.CreateCarPayload:
    properties:
      color:
        type: string
      company_id:
        type: integer
      make:
        type: string
      model:
        type: string
      price_per_day:
        type: number
      registration_no:
        type: string
      year:
        maximum: 2025
        minimum: 1886
        type: integer
    required:
    - color
    - company_id
    - make
    - model
    - price_per_day
    - registration_no
    - year
    type: object
...
  types.UpdateBookingPayload:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    type: object
  types.UpdateCarPayload:
    properties:
      color:
        type: string
      make:
        type: string
      model:
        type: string
      price_per_day:
        type: number
      registration_no:
        type: string
      year:
        maximum: 2025
        minimum: 1886
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /booking:
    post:
      consumes:
      - application/json
      description: Creates a new booking record
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.CreateBookingPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a booking
      tags:
      - Booking
  /booking/{id}:
    delete:
      description: Deletes a booking by its ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete booking by ID
      tags:
      - Booking
    get:
      description: Retrieves a booking based on the provided ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Booking'
      summary: Get booking by ID
      tags:
      - Booking
    put:
      consumes:
      - application/json
      description: Updates an existing booking with new data
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated booking data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.UpdateBookingPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update booking by ID
      tags:
      - Booking
  /booking/user/{id}:
    get:
      description: Retrieves all bookings for a user
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Booking'
            type: array
      summary: Get bookings by user ID
      tags:
      - Booking
  /car/{id}:
    delete:
      description: Deletes a car by its ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete car by ID
      tags:
      - Car
    get:
      description: Retrieves a car by its ID
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Car'
      summary: Get car by ID
      tags:
      - Car
    put:
      consumes:
      - application/json
      description: Updates a car's details
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated car data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.UpdateCarPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update car by ID
      tags:
      - Car
  /cars:
    get:
      description: Retrieves a list of cars with optional filters
      parameters:
      - description: Filters for car retrieval. eg. make[ct]=Mercedes&model[ct]=CLA&year=2022
        in: query
        name: filters
        type: string
      - description: sort for car retrieval, eg. id-asc
        in: query
        name: sort
        type: string
      - description: page number for car retrieval
        in: query
        name: page
        type: integer
      - description: number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Car'
            type: array
      summary: Get cars
      tags:
      - Car
    post:
      consumes:
      - application/json
      description: Creates a new car record
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.CreateCarPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a car
      tags:
      - Car
...
swagger: "2.0"